refactor(web): move charts related files under charts-view folder
