+++ apps/web/src/features/projects/views/project-impacts-page/charts-view/impacts/economic-balance/EconomicBalanceImpactCard.tsx
@@ -1,9 +1,9 @@
 import * as Highcharts from "highcharts";
 import HighchartsReact from "highcharts-react-official";
-import { baseColumnChartConfig } from "../../../shared/sharedChartConfig";
 import ImpactCard from "../../ImpactChartCard";
 
 import { ReconversionProjectImpacts } from "@/features/projects/domain/impacts.types";
+import { baseColumnChartConfig } from "@/features/projects/views/shared/sharedChartConfig.ts";
 import { formatNumberFr } from "@/shared/services/format-number/formatNumber";
 import { roundTo2Digits } from "@/shared/services/round-numbers/roundNumbers";
