+++ apps/web/src/features/projects/views/project-impacts-page/charts-view/impacts/environment/PermeableSurfaceImpactCard.tsx
@@ -1,8 +1,8 @@
 import * as Highcharts from "highcharts";
 import HighchartsReact from "highcharts-react-official";
-import { baseAreaChartConfig } from "../../../shared/sharedChartConfig";
 import ImpactCard from "../../ImpactChartCard";
 
+import { baseAreaChartConfig } from "@/features/projects/views/shared/sharedChartConfig.ts";
 import { formatNumberFr } from "@/shared/services/format-number/formatNumber";
 import { roundTo2Digits } from "@/shared/services/round-numbers/roundNumbers";
 import { SQUARE_METERS_HTML_SYMBOL } from "@/shared/views/components/SurfaceArea/SurfaceArea";
