+++ apps/web/src/features/projects/views/project-impacts-page/charts-view/impacts/environment/SoilsCarbonStorageImpactCard.tsx
@@ -1,10 +1,10 @@
 import * as Highcharts from "highcharts";
 import HighchartsReact from "highcharts-react-official";
-import { baseAreaChartConfig } from "../../../shared/sharedChartConfig";
 import ImpactCard from "../../ImpactChartCard";
-import { formatCO2Impact } from "../../list-view/formatImpactValue";
 
 import { ReconversionProjectImpacts } from "@/features/projects/domain/impacts.types";
+import { formatCO2Impact } from "@/features/projects/views/shared/formatImpactValue";
+import { baseAreaChartConfig } from "@/features/projects/views/shared/sharedChartConfig.ts";
 import { getColorForSoilType, SoilType } from "@/shared/domain/soils";
 import { getLabelForSoilType } from "@/shared/services/label-mapping/soilTypeLabelMapping";
 import { roundTo2Digits } from "@/shared/services/round-numbers/roundNumbers";
