+++ apps/web/src/features/projects/views/project-impacts-page/charts-view/impacts/socio-economic/SocioEconomicImpactsByActorChart.tsx
@@ -1,9 +1,10 @@
 import * as Highcharts from "highcharts";
 import HighchartsReact from "highcharts-react-official";
-import { baseColumnChartConfig } from "../../../shared/sharedChartConfig";
-import { getActorLabel, sumSocioEconomicImpactsByActor } from "./socioEconomicImpacts";
+import { sumSocioEconomicImpactsByActor } from "./socioEconomicImpacts";
 
 import { ReconversionProjectImpacts } from "@/features/projects/domain/impacts.types";
+import { baseColumnChartConfig } from "@/features/projects/views/shared/sharedChartConfig.ts";
+import { getActorLabel } from "@/features/projects/views/shared/socioEconomicLabels";
 import { formatNumberFr } from "@/shared/services/format-number/formatNumber";
 import { roundTo2Digits } from "@/shared/services/round-numbers/roundNumbers";
