+++ apps/web/src/features/projects/views/project-impacts-page/charts-view/impacts/socio-economic/SocioEconomicImpactsByCategoryChart.tsx
@@ -1,6 +1,5 @@
 import * as Highcharts from "highcharts";
 import HighchartsReact from "highcharts-react-official";
-import { baseColumnChartConfig } from "../../../shared/sharedChartConfig";
 import {
   getLabelForSocioEconomicImpactCategory,
   getNegativeSocioEconomicImpacts,
@@ -10,6 +9,7 @@ import {
 } from "./socioEconomicImpacts";
 
 import { ReconversionProjectImpacts } from "@/features/projects/domain/impacts.types";
+import { baseColumnChartConfig } from "@/features/projects/views/shared/sharedChartConfig.ts";
 import { formatNumberFr } from "@/shared/services/format-number/formatNumber";
 import { roundTo2Digits } from "@/shared/services/round-numbers/roundNumbers";
