+++ apps/web/src/features/projects/views/project-impacts-page/charts-view/impacts/socio-economic/SocioEconomicImpactsCard.tsx
